Fix gradfn attr bindings when saved variable is of an output (#53205)
Summary:
When saved variable is of an output, its grad_fn is not saved in SavedVariable, so it must be passed in during `unpack`.
Here, we can always pass in grad_fn (whether or not saved variable is an output) because it is ignored if the saved variable is not an output.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/53205

Reviewed By: gchanan, zhangguanheng66

Differential Revision: D26794365

Pulled By: soulitzer

fbshipit-source-id: e039baba20c364c4ab42ff99d0b242dd95c67fb3