[Pytorch Edge] Dont cache inflated bundled inputs (#55181)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55181

There can be a dramatic model size delta between saving a model after calling generate_bundled_inputs_for_* and saving before. This is due to the caching of the inflated tensor.

This increases latency when asking for the bundled inputs multiple times. I dont think this matters but it might for something like benchmarking?
ghstack-source-id: 125746773

Test Plan: unit tests.

Reviewed By: dreiss

Differential Revision: D27519487

fbshipit-source-id: 6ba22bff9c4e3a8d86c04627b7cbf47ca2d141b9